src/main/java/com/example/calculator/CalculatorServlet.java
package com.example.calculator;

import java.io.IOException;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

public class CalculatorServlet extends HttpServlet {
    protected void doPost(HttpServletRequest request, HttpServletResponse response)
            throws ServletException, IOException {
        int number1 = Integer.parseInt(request.getParameter("number1"));
        int number2 = Integer.parseInt(request.getParameter("number2"));
        String operator = request.getParameter("operator");
        int result = calculateResult(number1, number2, operator);
        System.out.println(result);
        request.setAttribute("result", result);
        request.getRequestDispatcher("/calculator.jsp").forward(request, response);
    }

    public int calculateResult(int number1, int number2, String operator) {
        switch (operator) {
            case "+":
                return number1 + number2;
            case "-":
                return number1 - number2;
            case "*":
                return number1 * number2;
            case "/":
                return number1 / number2;
            default:
                throw new IllegalArgumentException("Invalid operator");
        }
    }
}
